+++ src/Router/PublicRoute.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Layout from '../Layout/PrivateLayout';
-import Homepage from '../pages/homepage/Homepage';
+import Homepage from '../pages/homepage/homepage';
 import Login from '../pages/Login/Login';
 import PublicLayout from 'Layout/publicLayout';
 import SearchExistingCustomer from '@pages/SearchExistingCustomer/SearchExistingCustomer';
